feat: Enhance data filtering and display for Admins
This commit modifies the `CoursesList` and `StudentsList` pages to improve data filtering and display capabilities.

For administrators, the pages now correctly display all courses and students, respectively, regardless of selected program or batch. This is achieved by adding conditional logic to leverage `selectedCollegeId` for broader data retrieval when the user is an Admin.

Additionally, the logic for filtering courses and students based on the current user's role has been refined to ensure accurate data presentation for Program Co-ordinators and Teachers. The page titles and subtitles have also been updated to reflect these filtering changes.